update authorized wishlist customer

+++ src/api/store/wishlist/wishlist-item-delete.ts
@@ -15,35 +15,34 @@ export const deleteWishlistItem = async (req: MedusaRequest, res: MedusaResponse
         const manager: EntityManager = req.scope.resolve("manager");
         const wishListRepo = manager.withRepository(wishlistRepository);
 
-        if (req.user && req.user.customer_id) {
-
-            const wishlist = await wishListRepo.findOne({
-                where: {
-                    id: id
-                }
-            })
-
-            if (!wishlist) {
-                res.status(400).json({ message: "Wishlist Item Not Found" });
-                return
+        const wishlist = await wishListRepo.findOne({
+            where: {
+                id: id
             }
+        })
+        if (!wishlist) {
+            res.status(400).json({ message: "Wishlist Item Not Found" });
+            return
+        }
+
+
+        const wishlistNameRepository: typeof WishlistNameRepository =
+            req.scope.resolve("wishlistNameRepository");
+        const wishlistNameRepo = manager.withRepository(wishlistNameRepository);
+        const wishlistName = await wishlistNameRepo.findOne({
+            where: {
+                id: wishlist.wishlist_name_id
+            },
+            relations: ["wishlists"],
+        })
+        if (!wishlistName) {
+            res.status(404).json({ message: "Wishlist Name not found" });
+            return;
+        }
+
+        if (req.user && req.user.customer_id && req.user.customer_id === wishlistName.customer_id) {
+
             await wishListRepo.remove([wishlist])
-
-            const wishlistNameRepository: typeof WishlistNameRepository =
-                req.scope.resolve("wishlistNameRepository");
-            const wishlistNameRepo = manager.withRepository(wishlistNameRepository);
-            const wishlistName = await wishlistNameRepo.findOne({
-                where: {
-                    id: wishlist.wishlist_name_id
-                },
-                relations: ["wishlists"],
-            })
-
-            if (!wishlistName) {
-                res.status(404).json({ message: "Wishlist Name not found" });
-                return;
-            }
-
             res.json({ message: 'Wishlist Item Deleted' })
         } else {
             res.status(401).json({
